fix: implement T->T? implicit conversion and fix test issues
- Add T->T? implicit conversion in IR lowering for variable declarations
- Fix SafeCall test to use valid function instead of undefined to_string()
- Remove common test module to avoid cross-module dependency issues
- Inline test helper functions in each test file for better maintainability
- All 17 codegen tests now passing

Technical details:
- WrapOptional IR node now properly generated for T->T? conversions
- Generated code: 'let y: Int? = x' -> 'let y = Some(x)'
- Cleaner test structure without cross-module dependencies
