+++ crates/rive-semantic/tests/test_basic.rs
@@ -1,7 +1,27 @@
 //! Basic semantic analysis tests for variables, functions, and operations.
 
-mod common;
-use common::{should_fail, should_pass};
+use rive_core::Result;
+use rive_lexer::tokenize;
+use rive_parser::parse;
+use rive_semantic::analyze_with_registry;
+
+/// Helper to compile and analyze Rive source code.
+fn compile_and_analyze(source: &str) -> Result<()> {
+    let tokens = tokenize(source)?;
+    let (ast, type_registry) = parse(&tokens)?;
+    analyze_with_registry(&ast, type_registry)?;
+    Ok(())
+}
+
+/// Helper to check if source should pass.
+fn should_pass(source: &str) -> bool {
+    compile_and_analyze(source).is_ok()
+}
+
+/// Helper to check if source should fail.
+fn should_fail(source: &str) -> bool {
+    compile_and_analyze(source).is_err()
+}
 
 #[test]
 fn test_simple_variable_declaration() {
